+++ fdbserver/BulkDumpUtil.actor.cpp
@@ -105,7 +105,7 @@ std::string getBulkDumpJobRoot(const std::string& root, const UID& jobId) {
 	return appendToPath(root, jobId.toString());
 }
 
-std::pair<BulkDumpFileSet, BulkDumpFileSet> getLocalRemoteFileSetSetting(Version dumpVersion,
+std::pair<BulkLoadFileSet, BulkLoadFileSet> getLocalRemoteFileSetSetting(Version dumpVersion,
                                                                          const std::string& relativeFolder,
                                                                          const std::string& rootLocal,
                                                                          const std::string& rootRemote) {
@@ -113,8 +113,8 @@ std::pair<BulkDumpFileSet, BulkDumpFileSet> getLocalRemoteFileSetSetting(Version
 	const std::string manifestFileName = generateBulkDumpManifestFileName(dumpVersion);
 	const std::string dataFileName = generateBulkDumpDataFileName(dumpVersion);
 	const std::string byteSampleFileName = generateBulkDumpByteSampleFileName(dumpVersion);
-	BulkDumpFileSet fileSetLocal(rootLocal, relativeFolder, manifestFileName, dataFileName, byteSampleFileName);
-	BulkDumpFileSet fileSetRemote(rootRemote, relativeFolder, manifestFileName, dataFileName, byteSampleFileName);
+	BulkLoadFileSet fileSetLocal(rootLocal, relativeFolder, manifestFileName, dataFileName, byteSampleFileName);
+	BulkLoadFileSet fileSetRemote(rootRemote, relativeFolder, manifestFileName, dataFileName, byteSampleFileName);
 	return std::make_pair(fileSetLocal, fileSetRemote);
 }
 
@@ -172,14 +172,14 @@ void bulkDumpFileCopy(std::string fromFile, std::string toFile, size_t fileBytes
 }
 
 // Generate key-value data, byte sampling data, and manifest file given a range at a version with a certain bytes
-// Return BulkDumpManifest metadata (equivalent to content of the manifest file)
+// Return BulkLoadManifest metadata (equivalent to content of the manifest file)
 // TODO(BulkDump): can cause slow tasks, do the task in a separate thread in the future.
-BulkDumpManifest dumpDataFileToLocalDirectory(UID logId,
+BulkLoadManifest dumpDataFileToLocalDirectory(UID logId,
                                               const std::map<Key, Value>& sortedData,
                                               const std::map<Key, Value>& sortedSample,
-                                              const BulkDumpFileSet& localFileSetConfig,
-                                              const BulkDumpFileSet& remoteFileSetConfig,
-                                              const ByteSampleSetting& byteSampleSetting,
+                                              const BulkLoadFileSet& localFileSetConfig,
+                                              const BulkLoadFileSet& remoteFileSetConfig,
+                                              const BulkLoadByteSampleSetting& byteSampleSetting,
                                               Version dumpVersion,
                                               const KeyRange& dumpRange,
                                               int64_t dumpBytes) {
@@ -216,19 +216,19 @@ BulkDumpManifest dumpDataFileToLocalDirectory(UID logId,
 		ASSERT_WE_THINK(false);
 		throw retry();
 	}
-	BulkDumpFileSet fileSetRemote(remoteFileSetConfig.rootPath,
+	BulkLoadFileSet fileSetRemote(remoteFileSetConfig.rootPath,
 	                              remoteFileSetConfig.relativePath,
 	                              remoteFileSetConfig.manifestFileName,
 	                              containDataFile ? remoteFileSetConfig.dataFileName : "",
 	                              containByteSampleFile ? remoteFileSetConfig.byteSampleFileName : "");
-	BulkDumpManifest manifest(
+	BulkLoadManifest manifest(
 	    fileSetRemote, dumpRange.begin, dumpRange.end, dumpVersion, "", dumpBytes, byteSampleSetting);
 	writeStringToFile(abspath(localFiles.manifestFilePath), manifest.toString());
 	return manifest;
 }
 
 // Validate the invariant of filenames. Source is the file stored locally. Destination is the file going to move to.
-bool validateSourceDestinationFileSets(const BulkDumpFileSet& source, const BulkDumpFileSet& destination) {
+bool validateSourceDestinationFileSets(const BulkLoadFileSet& source, const BulkLoadFileSet& destination) {
 	// Manifest file must be present
 	if (source.manifestFileName.empty() || destination.manifestFileName.empty()) {
 		return false;
@@ -252,8 +252,8 @@ bool validateSourceDestinationFileSets(const BulkDumpFileSet& source, const Bulk
 }
 
 // Copy files between local file folders, used to mock blobstore in the test.
-void bulkDumpTransportCP_impl(BulkDumpFileSet sourceFileSet,
-                              BulkDumpFileSet destinationFileSet,
+void bulkDumpTransportCP_impl(BulkLoadFileSet sourceFileSet,
+                              BulkLoadFileSet destinationFileSet,
                               size_t fileBytesMax,
                               UID logId) {
 	BulkDumpFileFullPathSet localFiles(sourceFileSet);
@@ -275,8 +275,8 @@ void bulkDumpTransportCP_impl(BulkDumpFileSet sourceFileSet,
 }
 
 // Dump files to blobstore.
-ACTOR Future<Void> bulkDumpTransportBlobstore_impl(BulkDumpFileSet sourceFileSet,
-                                                   BulkDumpFileSet destinationFileSet,
+ACTOR Future<Void> bulkDumpTransportBlobstore_impl(BulkLoadFileSet sourceFileSet,
+                                                   BulkLoadFileSet destinationFileSet,
                                                    size_t fileBytesMax,
                                                    UID logId) {
 	// TODO(BulkDump): Make use of fileBytesMax
@@ -285,9 +285,9 @@ ACTOR Future<Void> bulkDumpTransportBlobstore_impl(BulkDumpFileSet sourceFileSet
 	return Void();
 }
 
-ACTOR Future<Void> uploadBulkDumpFileSet(BulkDumpTransportMethod transportMethod,
-                                         BulkDumpFileSet sourceFileSet,
-                                         BulkDumpFileSet destinationFileSet,
+ACTOR Future<Void> uploadBulkDumpFileSet(BulkLoadTransportMethod transportMethod,
+                                         BulkLoadFileSet sourceFileSet,
+                                         BulkLoadFileSet destinationFileSet,
                                          UID logId) {
 	// Validate file names of source and destination
 	if (!validateSourceDestinationFileSets(sourceFileSet, destinationFileSet)) {
@@ -298,10 +298,10 @@ ACTOR Future<Void> uploadBulkDumpFileSet(BulkDumpTransportMethod transportMethod
 		throw bulkdump_task_failed();
 	}
 	// Upload to blobstore or mock file copy
-	if (transportMethod == BulkDumpTransportMethod::BLOBSTORE) {
+	if (transportMethod == BulkLoadTransportMethod::BLOBSTORE) {
 		wait(bulkDumpTransportBlobstore_impl(
 		    sourceFileSet, destinationFileSet, SERVER_KNOBS->BULKLOAD_FILE_BYTES_MAX, logId));
-	} else if (transportMethod == BulkDumpTransportMethod::CP) {
+	} else if (transportMethod == BulkLoadTransportMethod::CP) {
 		bulkDumpTransportCP_impl(sourceFileSet, destinationFileSet, SERVER_KNOBS->BULKLOAD_FILE_BYTES_MAX, logId);
 	} else {
 		TraceEvent(SevError, "SSBulkDumpUploadFilesError", logId)
@@ -324,7 +324,7 @@ void generateBulkDumpJobManifestFile(const std::string& workFolder,
 	return;
 }
 
-ACTOR Future<Void> uploadBulkDumpJobManifestFile(BulkDumpTransportMethod transportMethod,
+ACTOR Future<Void> uploadBulkDumpJobManifestFile(BulkLoadTransportMethod transportMethod,
                                                  std::string localJobManifestFilePath,
                                                  std::string remoteFolder,
                                                  std::string remoteJobManifestFileName,
@@ -332,9 +332,9 @@ ACTOR Future<Void> uploadBulkDumpJobManifestFile(BulkDumpTransportMethod transpo
 	auto remoteJobManifestFilePath = appendToPath(remoteFolder, remoteJobManifestFileName);
 	TraceEvent(SevInfo, "UploadBulkDumpJobManifest", logId)
 	    .detail("RemoteJobManifestFilePath", remoteJobManifestFilePath);
-	if (transportMethod == BulkDumpTransportMethod::BLOBSTORE) {
+	if (transportMethod == BulkLoadTransportMethod::BLOBSTORE) {
 		wait(copyUpFile(localJobManifestFilePath, remoteJobManifestFilePath));
-	} else if (transportMethod == BulkDumpTransportMethod::CP) {
+	} else if (transportMethod == BulkLoadTransportMethod::CP) {
 		bulkDumpFileCopy(abspath(localJobManifestFilePath),
 		                 abspath(remoteJobManifestFilePath),
 		                 SERVER_KNOBS->BULKLOAD_FILE_BYTES_MAX,
@@ -398,7 +398,7 @@ ACTOR Future<Void> persistCompleteBulkDumpRange(Database cx, BulkDumpState bulkD
 	return Void();
 }
 
-std::string generateJobManifestFileContent(const std::map<Key, BulkDumpManifest>& manifests) {
+std::string generateJobManifestFileContent(const std::map<Key, BulkLoadManifest>& manifests) {
 	std::string root = "";
 	std::string manifestList;
 	for (const auto& [beginKey, manifest] : manifests) {
